Renamed sanitize to fix overload selection issues; Fixed Initial parse being empty; Added additional magic-variable recognitions

+++ server/src/variable_declaration.h
@@ -8,7 +8,7 @@
 
 struct variable_declaration
 {
-    int level;
+    size_t level;
     lsp::data::position position;
     std::string variable;
     std::vector<lsp::data::position> usages;
